% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a maps where the no of accidents are plotted on the state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A integer id of a state}

\item{year}{A integer which depicts the year}
}
\description{
This function generates a maps where the no of accidents are plotted on the 
state\code{year}
If wrong state code is given, then it will raise error
}
\section{Warning}{

If wrong state code is given, then it will raise error
}
\examples{
fars_map_state(state.num = 12, year = 2014)

}

