% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System data from csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string which is the name of the file}
}
\value{
This function returns a data frame
}
\description{
This function reads data from a .csv file, where the name of the file is supplied as
the argument of the function
If file does not exist in the current working directory, then the function will give error.
}

