% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return a dataframe which contains year and month column for a set of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list/vector containing years}
}
\value{
This function returns a dataframe which contains two columns, month and year
}
\description{
This function creates a dataframe for a set of years, which can be modified in
the code by changing \code{years}
If the dataset does not conatin the given year, then it will give warning
}
\examples{
fars_read_years(years = c(2013,2014,2015))

}

