% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return a dataframe which contains year and corresponding total entries, group by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list/vector containing years}
}
\value{
This function returns a dataframe which contains year and corresponding 
total entries, group by year and month
}
\description{
This function creates a dataframe for a set of years, which can be modified in
the code by changing \code{year}
}
\examples{
fars_summarize_years(years = c(2013,2014,2015))

}

