% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename from a given particular year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer which depicts the year}
}
\value{
This function returns a charcter string which is the name of the file
}
\description{
This function creates a filename for a particular year, which can be modified in
the code by changing \code{year}
}
\examples{
make_filename(year = 2014)

}

